robot-master/robot-label/requirements.txt
flask
docker
flask_pymongo
flask_cors
flask_socketio
gunicorn
gevent
gevent-websocket